Split vue files to html, css AND package.json improvements.
