+++ frontend/src/components/HelloWorld/HelloWorld.html
@@ -1,0 +1,3 @@
+<div class="hello">
+    <h1>{{ msg }}</h1>
+  </div>
