+++ frontend/src/components/HelloWorld/HelloWorld.vue
@@ -1,0 +1,12 @@
+<style scoped lang="scss" src="./HelloWorld.scss"></style>
+
+<template src="./HelloWorld.html"></template>
+
+<script>
+export default {
+  name: 'HelloWorld',
+  props: {
+    msg: String,
+  },
+};
+</script>
