+++ frontend/src/views/About/About.html
@@ -1,0 +1,3 @@
+<div class="about">
+    <h1>Routing Works!!</h1>
+</div>
