+++ frontend/src/views/About/About.vue
@@ -1,0 +1,7 @@
+<style scoped lang="scss" src="./About.scss"></style>
+<template src="./About.html"></template>
+<script>
+export default {
+  name: 'About',
+}
+</script>
