+++ frontend/src/views/Home/Home.html
@@ -1,0 +1,5 @@
+<div class="home">
+    <img alt="Vue logo" src="../assets/images/logo.png">
+    <HelloWorld msg="Hello Works!!"/>
+    <router-link to="/about">About</router-link>
+  </div>
